+++ CHANGELOG.md
@@ -8,7 +8,10 @@
 *  st-tooltip: Apply new style
 
 
-## 8.5.0 (March 07, 2018)
+**Others**
+
+*  st-pagination: Apply new style
+
 
 **New features:**